Separate directory library and binary
